% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tenmdata.R
\name{get_tenmdata}
\alias{get_tenmdata}
\title{Download time-specific bioclimatic variables at spatial resolution of 30" arcseconds 2.5', 5' and 10' arcminutes}
\usage{
get_tenmdata(res, w_dir = getwd())
}
\arguments{
\item{res}{resolution of the bioclimatic layers. Options are 30" arcseconds and 2.5', 5', 10' arcminutes}

\item{w_dir}{set the directory where layers will be written}
}
\value{
bioclimatic layers
}
\description{
Download time-specific bioclimatic variables at spatial resolution of 30" arcseconds 2.5', 5' and 10' arcminutes
}
\examples{
get_tenmdata(res = 2.5)
}
